# tpl/collections: Allow querify to accept a map argument