Closes #13131