@jmooring 2024-12-13
These are now equivalent:

```text
1. {{ collections.Querify "a" 1 "b" 2 }}
2. {{ collections.Querify (slice "a" 1 "b" 2) }}
3. {{ collections.Querify (dict "a" 1 "b" 2) }}
````

Benchmark testing indicates:
- Worse performance when passing a sequence of scalar values
- No change when passing a slice

I'm guessing that the performance degradation is primarily due to increased error checking and some deduplication of code. This function isn't widely used, so I suspect this won't be a problem for anyone.

```text
goos: linux
goarch: amd64
pkg: github.com/gohugoio/hugo/tpl/collections
cpu: Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz
                │     old     │                 new                 │
                │   sec/op    │    sec/op     vs base               │
Querify-12        670.6n ± 2%   905.4n ± 15%  +35.02% (p=0.002 n=6)
QuerifySlice-12   556.6n ± 5%   552.7n ±  4%        ~ (p=0.589 n=6)
QuerifyMap-12                   719.8n ±  4%
geomean           610.9n        711.5n        +15.79%

                │    old     │                 new                 │
                │    B/op    │    B/op     vs base                 │
Querify-12        128.0 ± 0%   224.0 ± 0%  +75.00% (p=0.002 n=6)
QuerifySlice-12   152.0 ± 0%   152.0 ± 0%        ~ (p=1.000 n=6) ¹
QuerifyMap-12                  128.0 ± 0%
geomean           139.5        163.3       +32.29%
¹ all samples are equal

                │    old     │                new                 │
                │ allocs/op  │ allocs/op   vs base                │
Querify-12        8.000 ± 0%   8.000 ± 0%       ~ (p=1.000 n=6) ¹
QuerifySlice-12   8.000 ± 0%   8.000 ± 0%       ~ (p=1.000 n=6) ¹
QuerifyMap-12                  7.000 ± 0%
geomean           8.000        7.652       +0.00%
¹ all samples are equal
```